mkdir implemented (was working) changed close to delete files changed assemble to delete chunks after assembling changed open close to implement link count so that assembled file is not deleted touch is working mkdir not working ls is working

+++ icfs/filesystem/file_object.py
@@ -1,8 +1,7 @@
+import os
 import time
 import uuid
-
-import os
-from stat import S_IFREG
+from stat import S_IFREG, S_IFDIR
 
 import constants
 from icfs.cloudapi.exceptions import CloudIOError
@@ -22,10 +21,11 @@ class FileObject:
         self.a_f_py_obj = None  # Python's File Object, populated after open
         self.cloud = cloud
 
-    def create(self, accounts):
+    def create(self, type, accounts):
         file_head_chunk_name = constants.HC_PREFIX + str(uuid.uuid4())
         self.head_chunk = HeadChunk(self.mpt, file_head_chunk_name, self.cloud,
                                     accounts)
+        self.head_chunk.type = type
         self.head_chunk.create()
 
     def create_root(self, accounts):
@@ -64,14 +64,21 @@ class FileObject:
         self.a_f_name = self.assemble()
         self.a_f_py_obj = open(os.path.join(self.mpt, self.a_f_name), flags)
 
-    def close(self):
+    def close(self, delete_local=False):
         if self.a_f_py_obj is not None:
             self.a_f_py_obj.close()
-            self.split_chunks()
             self.head_chunk.size = os.path.getsize(
                 os.path.join(self.mpt, self.a_f_name))
             self.head_chunk.write_file()
-            os.remove(os.path.join(self.mpt, self.a_f_name))
+            if delete_local:
+                if self.file_path != "/":
+                    os.remove(os.path.join(self.mpt, self.a_f_name))
+                    os.remove(
+                        os.path.join(self.mpt, self.head_chunk.chunk_meta.name))
+                    os.remove(os.path.join(self.mpt, self.head_chunk.name))
+                    self.head_chunk = None
+                    self.a_f_py_obj = None
+                    self.a_f_name = None
 
     # throws ICFS error
     def push(self):
@@ -85,20 +92,22 @@ class FileObject:
                     self.cloud.push(obj.name, acc)
                     acc_push_count += 1
                 except CloudIOError as cie:
-                    print "Error pushing into primary {}".format(cie.message)
+                    print "Error pushing into primary {} {}".format(obj.name,
+                                                                    cie.message)
             if acc_push_count < 1:
                 raise ICFSError("error while push")
 
     def assemble(self):
         chunks = self.head_chunk.chunk_meta.chunks
-        local_file_name = constants.LOCAL_ASSEMBLED_CHUNK + str(uuid.uuid4())
-        with open(os.path.join(self.mpt, local_file_name), "w") as of:
-            for chunk in chunks:
-                with open(os.path.join(self.mpt, chunk.name), "r") as chf:
-                    buf = chf.read(constants.CHUNK_SIZE)
-                    of.write(buf)
+        local_file_name = constants.LOCAL_ASSEMBLED_CHUNK + self.head_chunk.name
+        if not os.path.exists(os.path.join(self.mpt, local_file_name)):
+            with open(os.path.join(self.mpt, local_file_name), "w") as of:
+                for chunk in chunks:
+                    with open(os.path.join(self.mpt, chunk.name), "r") as chf:
+                        buf = chf.read(constants.CHUNK_SIZE)
+                        of.write(buf)
                     # TODO: Uncomment when Rsync is impl.
-                    # os.remove(os.path.join(self.mpt, chunk.name))
+                    os.remove(os.path.join(self.mpt, chunk.name))
         return local_file_name
 
     def write(self, data, offset):
@@ -106,6 +115,14 @@ class FileObject:
         self.a_f_py_obj.write(data)
         ret = self.a_f_py_obj.tell() - pos
         self.head_chunk.size += ret
+        self.split_chunks()
+        try:
+            self.push()  # Should be moved to write
+        except ICFSError as ie:
+            print
+            "Error in Pushing at FileSystem Layer. {}".format(ie.message)
+        for chunk in self.head_chunk.chunk_meta.chunks:
+            os.remove(os.path.join(self.mpt, chunk.name))
         return ret
 
     def read(self, length, offset):
@@ -116,17 +133,31 @@ class FileObject:
         self.head_chunk.fetch()
         self.head_chunk.load()
         now = time.time()
-        return dict(st_mode=(S_IFREG | 0o755), st_ctime=now, st_mtime=now,
-                    st_atime=now, st_nlink=1, st_size=self.head_chunk.size)
+        if self.head_chunk.type == constants.FILE:
+            return dict(st_mode=(S_IFREG | 0o755), st_ctime=now, st_mtime=now,
+                        st_atime=now, st_nlink=1, st_size=self.head_chunk.size)
+        else:
+            return dict(st_mode=(S_IFDIR | 0o755), st_ctime=now, st_mtime=now,
+                        st_atime=now, st_nlink=1, st_size=self.head_chunk.size)
 
     def split_chunks(self):
         f = open(os.path.join(self.mpt, self.a_f_name), 'r')
+        count = 0
         for chunk in self.head_chunk.chunk_meta.chunks:
             with open(os.path.join(self.mpt, chunk.name), 'w') as ch:
                 buf = f.read(constants.CHUNK_SIZE)
                 if buf == "":
                     break
                 ch.write(buf)
+            count += 1
+
+        while count < len(self.head_chunk.chunk_meta.chunks):
+            chunk = self.head_chunk.chunk_meta.chunks[count]
+            for client_id in chunk.accounts:
+                self.cloud.remove(chunk.name, client_id)
+            os.remove(os.path.join(self.mpt, chunk.name))
+            self.head_chunk.chunk_meta.chunks.remove(chunk)
+            count += 1
 
         buf = f.read(constants.CHUNK_SIZE)
         while buf != "":
